Update - CommentRepository 댓글 수 long 으로 변경 - 예외 클래스 명 변경 - PraiseComment 엔티티 네이밍 변경 - PraiseController 로그 레벨 변경 - PraiseController praise 조회 시 날짜 유효성 검사 추가 - Praise 엔티티 칭찬 보낸 사람 이름 sender 로 변경

+++ src/main/java/org/example/hugmeexp/domain/praise/repository/CommentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CommentRepository extends JpaRepository<PraiseComment, Long> {
-    int countByPraise(Praise praise);
+    long countByPraise(Praise praise);
 }
